Resolver: transmit the conflict without context
Both, OpenAI o1-mini and Claude Sonnet can process this input effectively and will respond with just the resolution. This resolves various issues with unexpected context lines (before and after) the actual conflict that we had encountered until now.

This also allows us to simplify the ResolverPatcher. Here we will only retain the handling of exact duplicate lines before and after the patch, which typically occurs for conflicts like:

```
<<<<<<< ours
	fun getSystemProperty(key: String): String? {
		return System.getProperty("${propertyPrefix}.$key")
=======
	fun getProperty(key: String, defaultValue: String): String {
		return System.getProperty("${propertyPrefix}.$key") ?: defaultValue
>>>>>>> theirs
	}
```